Tweaks to dangerous forms
- group deletion asserts requirement
- repos only need name confirmed
- only alert user prompt didn't match if they
  typed something

+++ gitbase/controllers/group.py
@@ -6,8 +6,11 @@ def group(group):
 
     if request.method == 'POST' and request.form.get('action') == 'group.delete':
         auth.assert_can('group.delete', group)
-        group.delete()
-        flash('Deleted group "%s"' % group.name)
-        return redirect(url_for('index'))
+        if request.form.get('user_accepted_danger'):
+            group.delete()
+            flash('Deleted group "%s"' % group.name)
+            return redirect(url_for('index'))
+        else:
+            flash('Javascript is required to delete groups.', 'danger')
 
     return render_template('group/group.haml', group=group)
